move CAR code to repo pkg
